// Tracker tool/nodejs-express-sequelize-mysql/Angular13Crud/src/app/models/tracker.model.ts
export class Tracker {

    id?: any;
    username?: string;
    userdescription?: string;
    userpublished?: boolean;
    createdAt?: string;
    updatedAt?: string;
    prioritytype?:string;
    worktype?:string; 
    membername?:string; 
    bankname?:string; 
    regionbank?:string; 
    solution?:string; 
    comment?:string
    
}